third_party/blink/web_tests/svg/animations/animate-elem-11-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect1.height.animVal.value is 210
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -250
PASS rect1.height.animVal.value is 143.4
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -183.4
PASS rect1.height.animVal.value is 143.4
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -183.4
PASS rect1.height.animVal.value is 76.7
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -116.7
PASS rect1.height.animVal.value is 76.7
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -116.7
PASS rect1.height.animVal.value is 10
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -50
PASS rect1.height.animVal.value is 10
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -50
PASS rect1.height.animVal.value is 10
PASS rect1.height.baseVal.value is 210
PASS getTransformToElement(rootSVGElement, rect2).e is -250
PASS getTransformToElement(rootSVGElement, rect2).f is -50
PASS successfullyParsed is true

TEST COMPLETE
